Refactoring et AppBar
